+++ rust-sdk/client/src/core_types/fusion_pool.rs
@@ -18,8 +18,6 @@ impl From<FusionPool> for FusionPoolFacade {
             tick_spacing: val.tick_spacing,
             fee_rate: val.fee_rate,
             protocol_fee_rate: val.protocol_fee_rate,
-            clp_reward_rate: val.clp_reward_rate,
-            order_protocol_fee_rate: val.order_protocol_fee_rate,
             liquidity: val.liquidity,
             sqrt_price: val.sqrt_price,
             tick_current_index: val.tick_current_index,