+++ ts-sdk/fusionamm-sdk/tests/limitOrder.test.ts
@@ -24,7 +24,7 @@ import {
   fetchFusionPool,
   fetchLimitOrder,
   fetchTickArray,
-  FusionPool,
+  FusionPool, getFusionPoolsConfigAddress,
   getLimitOrderAddress,
   getTickArrayAddress,
 } from "@crypticdot/fusionamm-client";
@@ -33,7 +33,12 @@ import { Account, Address, KeyPairSigner } from "@solana/kit";
 import assert from "assert";
 import { setupFusionPool } from "./utils/program";
 import { setupAtaTE, setupMintTE, setupMintTEFee } from "./utils/tokenExtensions";
-import { getTickArrayStartTickIndex, sqrtPriceToPrice } from "@crypticdot/fusionamm-core";
+import {
+  decreaseLimitOrderQuote,
+  getTickArrayStartTickIndex,
+  limitOrderQuoteByInputToken,
+  sqrtPriceToPrice,
+} from "@crypticdot/fusionamm-core";
 
 const mintTypes = new Map([
   ["A", setupMint],
@@ -56,7 +61,8 @@ const poolTypes = ["A-B", "A-TEA", "TEA-TEB", "A-TEFee"];
 describe("Limit Orders", () => {
   const tickSpacing = 64;
 
-  beforeAll(async () => {});
+  beforeAll(async () => {
+  });
 
   const fetchAndLogTickByTickIndex = async (fusionPool: Account<FusionPool>, tickIndex: number) => {
     const tickArrayStartIndex = getTickArrayStartTickIndex(tickIndex, fusionPool.data.tickSpacing);
@@ -249,15 +255,15 @@ describe("Limit Orders", () => {
       await testSwapExactInput({ poolAddress, inputAmount: 1_000_000n, mint: mintBAddress });
 
       fusionPool = await fetchFusionPool(rpc, poolAddress);
-      expect(fusionPool.data.protocolFeeOwedA).toEqual(750n);
-      expect(fusionPool.data.protocolFeeOwedB).toEqual(poolName == "A-TEFee" ? 741n : 750n);
+      expect(fusionPool.data.protocolFeeOwedA).toEqual(11n);
+      expect(fusionPool.data.protocolFeeOwedB).toEqual(poolName == "A-TEFee" ? 11n : 13n);
       expect(fusionPool.data.ordersTotalAmountA).toEqual(8000000n);
       expect(fusionPool.data.ordersTotalAmountB).toEqual(8000000n);
 
       expect(fusionPool.data.ordersFilledAmountA).toEqual(poolName == "A-TEFee" ? 4380687n : 4422472n);
       expect(fusionPool.data.ordersFilledAmountB).toEqual(4876385n);
-      expect(fusionPool.data.olpFeeOwedA).toEqual(754n);
-      expect(fusionPool.data.olpFeeOwedB).toEqual(poolName == "A-TEFee" ? 746n : 753n);
+      expect(fusionPool.data.olpFeeOwedA).toEqual(1493n);
+      expect(fusionPool.data.olpFeeOwedB).toEqual(poolName == "A-TEFee" ? 1476n : 1490n);
 
       for (const order of orders) {
         await testCloseLimitOrder({
@@ -266,8 +272,8 @@ describe("Limit Orders", () => {
       }
 
       fusionPool = await fetchFusionPool(rpc, poolAddress);
-      expect(fusionPool.data.protocolFeeOwedA).toEqual(750n);
-      expect(fusionPool.data.protocolFeeOwedB).toEqual(poolName == "A-TEFee" ? 741n : 750n);
+      expect(fusionPool.data.protocolFeeOwedA).toEqual(11n);
+      expect(fusionPool.data.protocolFeeOwedB).toEqual(poolName == "A-TEFee" ? 11n : 13n);
       expect(fusionPool.data.ordersTotalAmountA).toEqual(0n);
       expect(fusionPool.data.ordersTotalAmountB).toEqual(0n);
       expect(fusionPool.data.ordersFilledAmountA).toEqual(0n);
@@ -366,5 +372,75 @@ describe("Limit Orders", () => {
     expect(harvestIx.feesQuote.feeOwedB).equals(138n);
     expect(harvestIx.feesQuote.feeOwedA).equals(tokenAAfter.data.amount - tokenABefore.data.amount);
     expect(harvestIx.feesQuote.feeOwedB).equals(tokenBAfter.data.amount - tokenBBefore.data.amount);
+  });
+
+  it(`Quote and decrease limit order`, async () => {
+    const mintAName = "A";
+    const mintBName = "B";
+    const setupMintA = mintTypes.get(mintAName)!;
+    const setupMintB = mintTypes.get(mintBName)!;
+    const setupAtaA = ataTypes.get(mintAName)!;
+    const setupAtaB = ataTypes.get(mintBName)!;
+
+    const mintAAddress = await setupMintA();
+    const mintBAddress = await setupMintB();
+    const mintA = await fetchMint(rpc, mintAAddress);
+    const mintB = await fetchMint(rpc, mintBAddress);
+    const ataAAddress = await setupAtaA(mintAAddress, { amount: 100_000_000n });
+    const ataBAddress = await setupAtaB(mintBAddress, { amount: 100_000_000n });
+    const poolAddress = await setupFusionPool(mintAAddress, mintBAddress, tickSpacing);
+
+    const limitOrdersArgs = [
+      { amount: 1_000_000n, priceOffset: -0.06, aToB: false }, // 1st
+      { amount: 1_000_000n, priceOffset: -0.1, aToB: false }, // 2nd
+    ];
+
+    const orders = [];
+    let fusionPool = await fetchFusionPool(rpc, poolAddress);
+    let currentPrice = sqrtPriceToPrice(fusionPool.data.sqrtPrice, mintA.data.decimals, mintB.data.decimals);
+
+    for (const args of limitOrdersArgs) {
+      orders.push(
+        await testOpenLimitOrder({
+          ...args,
+          priceOrTickIndex: { price: currentPrice + args.priceOffset },
+          poolAddress,
+          signer,
+        }),
+      );
+    }
+
+    const limitOrder = orders[0].data;
+
+    // Quote the limit order output
+    const quotedAmountOut = limitOrderQuoteByInputToken(orders[0].data.amount, orders[0].data.aToB, orders[0].data.tickIndex, fusionPool.data);
+    // The actual limit order output will be 1066409.
+    // It happens because the quote function has different math. A small error is fine.
+    expect(quotedAmountOut).toEqual(1066405n);
+
+    // The 1st order will be fulfilled.
+    await testSwapExactInput({ poolAddress, inputAmount: 1_500_000n, mint: mintAAddress });
+
+    fusionPool = await fetchFusionPool(rpc, poolAddress);
+    const startTickIndex = getTickArrayStartTickIndex(limitOrder.tickIndex, fusionPool.data.tickSpacing);
+    const tickArrayAddress = await getTickArrayAddress(fusionPool.address, startTickIndex);
+    const tickArray = await fetchTickArray(rpc, tickArrayAddress[0]);
+
+    // Decrease Limit Order Quote
+    const tick = tickArray.data.ticks[(limitOrder.tickIndex - startTickIndex) / fusionPool.data.tickSpacing];
+    const decreaseQuote = decreaseLimitOrderQuote(fusionPool.data, limitOrder, tick, limitOrder.amount);
+    expect(decreaseQuote.amountOutA).toEqual(1066409n);
+    expect(decreaseQuote.amountOutB).toEqual(0n);
+
+    // Execute the decrease order instruction
+    const tokenBeforeA = await fetchToken(rpc, ataAAddress);
+    const tokenBeforeB = await fetchToken(rpc, ataBAddress);
+    await testCloseLimitOrder({
+      limitOrderMint: orders[0].data.limitOrderMint,
+    });
+    const tokenAfterA = await fetchToken(rpc, ataAAddress);
+    const tokenAfterB = await fetchToken(rpc, ataBAddress);
+    expect(tokenAfterA.data.amount - tokenBeforeA.data.amount).toEqual(1066409n);
+    expect(tokenAfterB.data.amount - tokenBeforeB.data.amount).toEqual(0n);
   });
 });