+++ rust-sdk/core/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @crypticdot/fusionamm-rust-core
 
+## 1.0.66
+
+### Patch Changes
+
+- 3020148: Simplified limit order rewards calculation math
+
+## 1.0.65
+
+### Patch Changes
+
+- 9606b26: Added limit_order_reward_by_output_token()
+
+## 1.0.64
+
+## 1.0.63
+
+### Patch Changes
+
+- 74e1d40: Fixed incorrect math in limit_order_quote_by_input_token()
+
 ## 1.0.62
 
 ### Patch Changes